Crate buck2_downward_api_proto

Protobufs for ineteracting with Buck’s DownwardApi over GPRC. This isn’t the protocol Buck v1 speaks, where the DownwardApi is accessed over named pipes with serialized JSON payloads. This is a different way to make the same calls.

Modules

downward_api_client

Generated client implementations.

downward_api_server

Generated server implementations.

event

Nested message and enum types in Event.

log_level

Nested message and enum types in LogLevel.

Structs

ConsoleRequest

Empty

Event

ExternalEventRequest

LogLevel

LogRequest